+++ server-spi-private/src/main/java/org/keycloak/models/UserModelDefaultMethods.java
